// packages/db/generated/schemas/objects/TokenWhereInput.schema.ts
const tokenwhereinputSchema = z.object({
  AND: z.union([z.lazy(() => TokenWhereInputObjectSchema), z.lazy(() => TokenWhereInputObjectSchema).array()]).optional(),
  OR: z.lazy(() => TokenWhereInputObjectSchema).array().optional(),
  NOT: z.union([z.lazy(() => TokenWhereInputObjectSchema), z.lazy(() => TokenWhereInputObjectSchema).array()]).optional(),
  id: z.union([z.lazy(() => StringFilterObjectSchema), z.string().max(36)]).optional(),
  createdAt: z.union([z.lazy(() => DateTimeFilterObjectSchema), z.coerce.date()]).optional(),
  updatedAt: z.union([z.lazy(() => DateTimeFilterObjectSchema), z.coerce.date()]).optional(),
  name: z.union([z.lazy(() => StringFilterObjectSchema), z.string()]).optional(),
  keyHash: z.union([z.lazy(() => StringFilterObjectSchema), z.string()]).optional(),
  keyPrefix: z.union([z.lazy(() => StringFilterObjectSchema), z.string()]).optional(),
  ownerModel: z.union([z.lazy(() => EnumTokenOwnerModelFilterObjectSchema), TokenOwnerModelSchema]).optional(),
  userId: z.union([z.lazy(() => StringNullableFilterObjectSchema), z.string().max(36)]).optional().nullable(),
  organizationId: z.union([z.lazy(() => StringNullableFilterObjectSchema), z.string().max(36)]).optional().nullable(),
  role: z.union([z.lazy(() => EnumOrganizationRoleFilterObjectSchema), OrganizationRoleSchema]).optional(),
  entitlements: z.lazy(() => JsonNullableFilterObjectSchema).optional(),
  expiresAt: z.union([z.lazy(() => DateTimeNullableFilterObjectSchema), z.coerce.date()]).optional().nullable(),
  lastUsedAt: z.union([z.lazy(() => DateTimeNullableFilterObjectSchema), z.coerce.date()]).optional().nullable(),
  isActive: z.union([z.lazy(() => BoolFilterObjectSchema), z.boolean()]).optional(),
  user: z.union([z.lazy(() => UserNullableScalarRelationFilterObjectSchema), z.lazy(() => UserWhereInputObjectSchema)]).optional(),
  organization: z.union([z.lazy(() => OrganizationNullableScalarRelationFilterObjectSchema), z.lazy(() => OrganizationWhereInputObjectSchema)]).optional(),
  organizationUser: z.union([z.lazy(() => OrganizationUserNullableScalarRelationFilterObjectSchema), z.lazy(() => OrganizationUserWhereInputObjectSchema)]).optional()
}).strict();
export const TokenWhereInputObjectSchema: z.ZodType<Prisma.TokenWhereInput> = tokenwhereinputSchema as unknown as z.ZodType<Prisma.TokenWhereInput>;